Digitalized_Student_Management_System/flask/requirements.txt
# # Flask and its extensions
# Flask==2.3.2
# flask-cors==3.0.10

# # Environment variable management
# python-dotenv==1.0.0

# # MongoDB client
# pymongo==4.5.0

# # Cloudinary for file uploads
# cloudinary==1.30.0

# # InsightFace for facial recognition
# insightface==0.7.3

# # ONNX Runtime (required by InsightFace)
# onnxruntime==1.16.0

# # Computer vision libraries
# opencv-python-headless==4.8.1.78
# Pillow==10.0.1
# numpy==1.26.0

# # HTTP requests
# requests==2.31.0

opencv-python-headless==4.8.0.74  # Avoid the full OpenCV library to save space
numpy==1.21.0  # A lighter version compatible with insightface
insightface==0.7  # Choose a stable version
requests==2.31.0
python-dotenv==1.0.0
Flask==2.2.5
flask-cors==3.0.10
onnxruntime==1.16.0  # Use only for inference
pillow==9.4.0
pymongo==4.5.0
cloudinary==1.33.0